+++ _boards/frdmmcxe247/ota_examples/ota_mcuboot_basic/reconfig.cmake
@@ -1,0 +1,44 @@
+# Add additional configuration
+mcux_add_iar_configuration(
+    CX "--diag_suppress=Pe257,Pa039"
+)
+mcux_add_mdk_configuration(
+    LD "--diag_suppress=L6329W"
+)
+
+
+# Add or remove Linker File Configurations
+
+mcux_remove_iar_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${device_root}/MCX/MCXE/MCXE247/iar/MCXE247_flash.icf
+)
+
+mcux_remove_armgcc_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${device_root}/MCX/MCXE/MCXE247/gcc/MCXE247_flash.ld
+)
+
+mcux_remove_mdk_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${device_root}/MCX/MCXE/MCXE247/arm/MCXE247_flash.csf
+)
+
+mcux_add_iar_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${board_root}/${board}/ota_examples/linker/MCXE247_flash.icf
+)
+mcux_add_mdk_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${board_root}/${board}/ota_examples/linker/MCXE247_flash.scf
+)
+mcux_add_armgcc_linker_script(
+    BASE_PATH ${SdkRootDirPath}
+    TARGETS debug release
+    LINKER ${board_root}/${board}/ota_examples/linker/MCXE247_flash.ld
+)